// src/shared/types/product.ts
export type Product = {
  codigo: string;              // Código de identificação do produto
  nome: string;                // Nome do produto
  referencia: string;          // Código de referencia do produto
  codigo_categoria: string;    // Código da categoria do produto
  imagem: string;              // Link para imagem do produto
  preco: string;               // Preço do produto (string)
  descricao: string;           // Descrição do produto
};